+++ src/plugins/plugin.rs
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Manuel Quarneti <mq1@ik.me>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::plugins::tool::ToolMeta;
 use serde::Deserialize;
 use std::path::PathBuf;
 
@@ -21,4 +20,12 @@ pub struct PluginMeta {
     pub license: String,
     pub runs_on: Vec<String>,
     pub tools: Vec<ToolMeta>,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct ToolMeta {
+    pub name: String,
+    pub description: String,
+    pub icon: String,
+    pub group: String,
 }
